refactor: extract form validation into reusable helper function
